requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.2
WTForms==3.2.2
email-validator==2.2.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
